Basic Extension Plugin Working
this current state will work by specifying key:value pairs in a JSON file. It will  search for a registered extension and install it if the name matches. At this point, rexray is the only that has been tested. Will need to develop this further by having multi-nested params in the JSON file. Will have 2 different categories. First is env vars, second is files. 3rd may be commands if it's found to be necessary

Signed-off-by: Kendrick Coleman <kendrickcoleman@gmail.com>