completed file transfer using scp
completed part 2 of the file transfer process using the docker-machine scp command to move files from any host to another and even locally. Changed the log messages to use the name to uppercase version so it's easily readable during output. Changed the log to warn instead of debug if an extension isn't found in the json file

Signed-off-by: Kendrick Coleman <kendrickcoleman@gmail.com>